Move network related commands to cmd package
