+++ cmd/network_darwin.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import "github.com/vishvananda/netlink"
