+++ ops.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path"
 	"path/filepath"
 	"runtime"
@@ -455,33 +454,6 @@ func imageCommandHandler(cmd *cobra.Command, args []string) {
 	}
 }
 
-func runningAsRoot() bool {
-	cmd := exec.Command("id", "-u")
-	output, _ := cmd.Output()
-	i, _ := strconv.Atoi(string(output[:len(output)-1]))
-	return i == 0
-}
-
-func netSetupCommandHandler(cmd *cobra.Command, args []string) {
-	if !runningAsRoot() {
-		fmt.Println("net command needs root permission")
-		return
-	}
-	if err := setupBridgeNetwork(); err != nil {
-		panic(err)
-	}
-}
-
-func netResetCommandHandler(cmd *cobra.Command, args []string) {
-	if !runningAsRoot() {
-		fmt.Println("net command needs root permission")
-		return
-	}
-	if err := resetBridgeNetwork(); err != nil {
-		panic(err)
-	}
-}
-
 // merge userconfig to package config, user config takes precedence
 func mergeConfigs(pkgConfig *api.Config, usrConfig *api.Config) *api.Config {
 
@@ -694,25 +666,6 @@ func main() {
 	cmdRun.PersistentFlags().StringVarP(&manifestName, "manifest-name", "m", "", "save manifest to file")
 	cmdRun.PersistentFlags().BoolVarP(&accel, "accel", "x", false, "use cpu virtualization extension")
 
-	var cmdNetSetup = &cobra.Command{
-		Use:   "setup",
-		Short: "Setup bridged network",
-		Run:   netSetupCommandHandler,
-	}
-
-	var cmdNetReset = &cobra.Command{
-		Use:   "reset",
-		Short: "Reset bridged network",
-		Run:   netResetCommandHandler,
-	}
-
-	var cmdNet = &cobra.Command{
-		Use:       "net",
-		Args:      cobra.OnlyValidArgs,
-		ValidArgs: []string{"setup", "reset"},
-		Short:     "Configure bridge network",
-	}
-
 	var cmdBuild = &cobra.Command{
 		Use:   "build [ELF file]",
 		Short: "Build an image from ELF",
@@ -810,14 +763,12 @@ func main() {
 
 	var rootCmd = &cobra.Command{Use: "ops"}
 	rootCmd.AddCommand(cmdRun)
-	cmdNet.AddCommand(cmdNetReset)
-	cmdNet.AddCommand(cmdNetSetup)
-	rootCmd.AddCommand(cmdNet)
+	rootCmd.AddCommand(cmd.NetCommands())
 	rootCmd.AddCommand(cmdBuild)
 	rootCmd.AddCommand(cmdPrintConfig)
 	rootCmd.AddCommand(cmdVersion)
 	rootCmd.AddCommand(cmdUpdate)
-	rootCmd.AddCommand(cmd.PackageCommand())
+	rootCmd.AddCommand(cmd.PackageCommands())
 	rootCmd.AddCommand(cmdLoadPackage)
 	rootCmd.AddCommand(cmdGetPackage)
 	rootCmd.AddCommand(cmdInstance)
